Moved common helper to more logical location
