+++ routes/order.js
@@ -1,5 +1,5 @@
 const express = require('express');
-const common = require('./common');
+const common = require('../lib/common');
 const router = express.Router();
 
 // Show orders
